Move some extrinsic operations to common
