+++ e2e-tests/src/test/treasury.rs
@@ -1,9 +1,10 @@
-use crate::accounts::{accounts_from_seeds, get_free_balance, get_sudo};
-use crate::config::Config;
-use crate::fee::get_tx_fee_info;
-use crate::transfer::{setup_for_transfer, transfer};
-use crate::waiting::wait_for_event;
-use crate::Connection;
+use crate::{
+    accounts::{accounts_from_seeds, get_free_balance, get_sudo},
+    config::Config,
+    fee::get_tx_fee_info,
+    transfer::{setup_for_transfer, transfer},
+    wait_for_event, Connection,
+};
 use codec::{Compact, Decode};
 use common::create_connection;
 use frame_support::PalletId;
