+++ e2e-tests/src/test/validators_change.rs
@@ -7,8 +7,8 @@ use substrate_api_client::AccountId;
 use crate::accounts::{accounts_from_seeds, get_sudo};
 use crate::config::Config;
 use crate::session::send_change_members;
-use crate::waiting::{wait_for_event, wait_for_finalized_block};
-use crate::Header;
+use crate::waiting::wait_for_finalized_block;
+use crate::{wait_for_event, Header};
 
 pub fn change_validators(config: &Config) -> anyhow::Result<()> {
     let Config {
